도메인, 컨트롤러 분리 & strategy 의존성 제거
